requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.11.12
channels==4.0.0
channels-redis==4.2.1
charset-normalizer==3.4.4
dj-database-url==3.1.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
dotenv==0.9.9
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
PyJWT==2.10.1
PyYAML==6.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
psycopg2-binary==2.9.9
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
django-jazzmin==3.0.1
django-cors-headers==4.9.0
